Добавление изменений и репозиториев

+++ models/expense.py
@@ -7,7 +7,7 @@ class Expense(Base):
     id = Column(Integer, primary_key=True, index=True)
     trip_id = Column(Integer, ForeignKey("trips.id"))
     user_id_pay = Column(Integer, ForeignKey("users.id"))
-    amount = Column(Integer, nullable=False)
+    amount = Column(Float, nullable=False)
     date = Column(Date, nullable=False)
     type_of_expense = Column(Integer, ForeignKey("expenseTypes.id"))
     description = Column(String, nullable=True)
