+++ models/tripBudgetCategory.py
@@ -1,5 +1,5 @@
 from database import Base
-from sqlalchemy import Column, Integer, ForeignKey
+from sqlalchemy import Column, Integer, ForeignKey, Float
 
 class TripBudgetCategory(Base):
     __tablename__ = "tripBudgetCategories"
@@ -7,4 +7,4 @@ class TripBudgetCategory(Base):
     id = Column(Integer, primary_key=True, index=True)
     trip_id = Column(Integer, ForeignKey("trips.id"))
     expense_type_id = Column(Integer, ForeignKey("expenseTypes.id"))
-    planned_amount = Column(Integer, nullable=False)
+    planned_amount = Column(Float, nullable=False)
